requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
anyio==4.8.0
APScheduler==3.6.3
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==21.2.0
base58==2.1.1
bcc==0.18.0
bech32==1.2.0
bip-utils==2.9.3
bip32utils==0.3.post3
bitarray==2.9.3
bitcoinlib==0.7.2
blinker==1.4
Brlapi==0.8.3
btclib @ git+https://github.com/btclib-org/btclib.git@f05e77e4ca807893d63df3f1535563554b3e22a3
btclib_libsecp256k1==0.4.0
cachetools==4.2.2
cbor2==5.6.5
certifi==2024.12.14
cffi==1.17.1
chacha20poly1305==0.0.3
chardet==4.0.0
charset-normalizer==3.4.1
cheroot==10.0.1
chrome-gnome-shell==0.0.0
click==8.1.8
coincurve==20.0.0
colorama==0.4.4
command-not-found==0.3
construct==2.10.70
construct-classes==0.1.2
crcmod==1.7
crypto-licensing==3.3.2
cryptography==3.4.8
cupshelpers==1.0
cytoolz==1.0.1
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
diceware==0.9.6
distlib==0.3.9
distro==1.7.0
dnspython==2.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.19.0
ed25519-blake2b==1.4.1
embit @ file:///home/roger/workarea/embit
eth-account==0.8.0
eth-hash==0.7.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.2
eth_abi==5.2.0
evdev==1.4.0
exceptiongroup==1.2.2
fastecdsa==3.0.0
filelock==3.17.0
fonttools==4.55.4
fpdf2==2.8.2
frozenlist==1.5.0
future==1.0.0
fuzzywuzzy==0.18.0
greenlet==3.1.1
greppy==0.1.3
h11==0.14.0
hdwallet==2.2.1
hdwallet-slip39==2.3.0
hexbytes==0.3.1
hidpidaemon==18.4.6
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
idna==3.10
importlib-metadata==4.6.4
iotop==0.6
jaraco.functools==4.1.0
jeepney==0.7.1
jsonschema==3.2.0
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libusb1==3.2.0
libvirt-python==8.0.0
louis==3.20.0
macaroonbakery==1.3.1
mnemonic==0.21
more-itertools==10.6.0
multidict==6.1.0
netaddr==0.8.0
netifaces==0.11.0
numpy==2.2.2
oauthlib==3.2.0
parsimonious==0.10.0
petname==2.6
pillow==11.1.0
platformdirs==4.3.6
pop-transition==1.1.2
propcache==0.3.0
protobuf==3.12.4
py-sr25519-bindings==0.2.1
pycairo==1.20.1
pycoin==0.92.20241201
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.11.0
pycups==2.0.1
pydbus==0.6.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySimpleGUI==4.70.1
python-apt==2.4.0+ubuntu4
python-bitcoinlib==0.11.0
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-Levenshtein==0.12.2
python-telegram-bot==13.15
python-xlib==0.29
pytz==2024.2
pyxdg==0.27
PyYAML==5.4.1
qrcode==8.0
regex==2024.11.6
repolib==2.2.1
repoman==1.4.0
requests==2.32.3
rlp==3.0.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
shamir-mnemonic==0.3.0
shamir-mnemonic-slip39==0.4.1
six==1.17.0
slip39==12.1.0
sniffio==1.3.1
SQLAlchemy==2.0.37
ssh-import-id==5.11
system76driver==22.4.0
systemd-python==234
tabulate-slip39==0.10.4
texttable==1.6.4
toolz==1.0.0
tornado==6.1
trezor==0.13.9
typing_extensions==4.12.2
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==2.3.0
virtualenv==20.29.1
wadllib==1.3.6
web.py==0.62
websocket-client==1.2.3
wsgilog==0.3.1
xdg==5
xkit==0.0.0
yarl==1.18.3
zipp==1.0.0
